+++ OMG.PSUtilities.AzureDevOps/Public/Complete-PSUADOPullRequest.ps1
@@ -99,7 +99,6 @@ function Complete-PSUADOPullRequest {
         [string]$Organization = $env:ORGANIZATION,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
         [string]$PAT = $env:PAT
     )