+++ OMG.PSUtilities.AzureDevOps/Public/Get-PSUADORepoBranchList.ps1
@@ -75,7 +75,6 @@ function Get-PSUADORepoBranchList {
         [string]$Organization = $env:ORGANIZATION,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
         [string]$PAT = $env:PAT
     )