+++ OMG.PSUtilities.AzureDevOps/Public/Get-PSUADOVariableGroup.ps1
@@ -87,7 +87,6 @@ function Get-PSUADOVariableGroup {
         [string]$Organization = $env:ORGANIZATION,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
         [string]$PAT = $env:PAT
     )