+++ OMG.PSUtilities.AzureDevOps/Public/Get-PSUADOWorkItemStates.ps1
@@ -66,7 +66,6 @@ function Get-PSUADOWorkItemStates {
         [string]$Organization = $env:ORGANIZATION,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
         [string]$PAT = $env:PAT
     )