+++ OMG.PSUtilities.AzureDevOps/Public/New-PSUADOUserStory.ps1
@@ -119,7 +119,6 @@ function New-PSUADOUserStory {
         [string]$Organization = $env:ORGANIZATION,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
         [string]$PAT = $env:PAT
     )