+++ OMG.PSUtilities.AzureDevOps/Public/Set-PSUADOBug.ps1
@@ -140,7 +140,6 @@ function Set-PSUADOBug {
         [string]$Organization = $env:ORGANIZATION,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
         [string]$PAT = $env:PAT
     )